projects/daas-data-lake-sub-dataflow-bigquery-develop/daas-data-lake-sub-dataflow-bigquery-develop/terragenesis/src/functions/datalake-availability/requirements.txt
google-cloud-monitoring
google-cloud-pubsub
google-cloud-bigquery
google-cloud-secret-manager
PyYAML==6.0